Exception handling

+++ src/main/java/com/biazon/dscommerce/services/ProductService.java
@@ -1,14 +1,17 @@
 package com.biazon.dscommerce.services;
 
 import org.springframework.beans.BeanUtils;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.biazon.dscommerce.entiteis.Product;
 import com.biazon.dscommerce.entiteis.dtos.ProductInsertDTO;
 import com.biazon.dscommerce.entiteis.dtos.ProductResponseDTO;
+import com.biazon.dscommerce.exceptions.DatabaseException;
 import com.biazon.dscommerce.exceptions.ResourceNotFoundException;
 import com.biazon.dscommerce.repositories.ProductRepository;
 
@@ -30,9 +33,9 @@ public class ProductService {
 
 	@Transactional(readOnly = true)
 	public ProductResponseDTO getOneProduct(Long id) {
-		return new ProductResponseDTO(productRepository.findById(id).orElseThrow(
-				() -> new ResourceNotFoundException("Produto não encontrado")));
-		
+		return new ProductResponseDTO(productRepository.findById(id)
+				.orElseThrow(() -> new ResourceNotFoundException("Produto não encontrado")));
+
 	}
 
 	@Transactional
@@ -45,15 +48,26 @@ public class ProductService {
 
 	@Transactional
 	public ProductResponseDTO updateProducts(Long id, ProductInsertDTO dto) {
-		Product product = productRepository.getReferenceById(id);
-		BeanUtils.copyProperties(dto, product);
-
-		return new ProductResponseDTO(productRepository.save(product));
+		try {
+			Product product = productRepository.getReferenceById(id);
+			BeanUtils.copyProperties(dto, product);
+			return new ProductResponseDTO(productRepository.save(product));
+		} catch (ResourceNotFoundException e) {
+			throw new ResourceNotFoundException("Produto não encontrado");
+		}
 	}
-	
-	@Transactional
+
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void deletProduct(Long id) {
-		productRepository.deleteById(id);
+		if(!productRepository.existsById(id)) {
+			throw new ResourceNotFoundException("Produto não encontrado");
+		}
+		try {
+			productRepository.deleteById(id);
+		} catch (DataIntegrityViolationException e) {
+			throw new DatabaseException("Falha de integridade refencial");
+		}
+		
 	}
 
 }
